+++ Source/Cells/OptionViewControllerCell.swift
@@ -8,7 +8,7 @@ public struct OptionViewControllerCellModel {
 	var selectedOptionRow: OptionRowModel? = nil
 
 	var valueDidChange: (OptionRowModel?) -> Void = { (value: OptionRowModel?) in
-		SwiftyFormLog("value \(value)")
+		SwiftyFormLog("value \(String(describing: value))")
 	}
 }
 
@@ -41,12 +41,12 @@ public class OptionViewControllerCell: UITableViewCell, SelectRowDelegate {
 	
 	fileprivate func updateValue() {
 		let s = humanReadableValue()
-		SwiftyFormLog("update value \(s)")
+		SwiftyFormLog("update value \(String(describing: s))")
 		detailTextLabel?.text = s
 	}
 	
 	public func setSelectedOptionRowWithoutPropagation(_ option: OptionRowModel?) {
-		SwiftyFormLog("set selected option: \(option?.title) \(option?.identifier)")
+		SwiftyFormLog("set selected option: \(String(describing: option?.title)) \(String(describing: option?.identifier))")
 		
 		selectedOptionRow = option
 		updateValue()