+++ Source/Form/FormViewController.swift
@@ -10,25 +10,30 @@ open class FormViewController: UIViewController {
 		super.init(nibName: nil, bundle: nil)
 	}
 	
-	required public init(coder aDecoder: NSCoder) {
+	required public init?(coder aDecoder: NSCoder) {
 		SwiftyFormLog("super init")
-		super.init(nibName: nil, bundle: nil)
+		super.init(coder: aDecoder)
 	}
 
 	override open func loadView() {
 		SwiftyFormLog("super loadview")
 		view = tableView
-		
 		keyboardHandler = KeyboardHandler(tableView: tableView)
-		
+		populateAndSetup()
+	}
+	
+	open func populateAndSetup() {
 		populate(formBuilder)
 		title = formBuilder.navigationTitle
-		
 		dataSource = formBuilder.result(self)
-		self.tableView.dataSource = dataSource
-		self.tableView.delegate = dataSource
-		
-		//debugPrint(dataSource!)
+		tableView.dataSource = dataSource
+		tableView.delegate = dataSource
+	}
+
+	open func reloadForm() {
+		formBuilder.removeAll()
+		populateAndSetup()
+		tableView.reloadData()
 	}
 
 	open func populate(_ builder: FormBuilder) {